adding jwt claims

+++ resources/item.py
@@ -1,26 +1,26 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.item import ItemModel
+
+_item_parser = reqparse.RequestParser()
+_item_parser.add_argument('name',
+                          type=str,
+                          required=True,
+                          help="This field cannot be left blank."
+                          )
+_item_parser.add_argument('price',
+                          type=float,
+                          required=True,
+                          help="This field cannot be left blank."
+                          )
+_item_parser.add_argument('store_id',
+                          type=int,
+                          required=True,
+                          help="Every item needs a store id."
+                          )
 
 
 class Item(Resource):
-    parser = reqparse.RequestParser()
-    parser.add_argument('name',
-                        type=str,
-                        required=True,
-                        help="This field cannot be left blank."
-                        )
-    parser.add_argument('price',
-                        type=float,
-                        required=True,
-                        help="This field cannot be left blank."
-                        )
-    parser.add_argument('store_id',
-                        type=int,
-                        required=True,
-                        help="Every item needs a store id."
-                        )
-
     @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
@@ -28,7 +28,11 @@ class Item(Resource):
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -36,7 +40,7 @@ class Item(Resource):
         return {'message': 'Item deleted'}
 
     def put(self, name):
-        data = Item.parser.parse_args()
+        data = _item_parser.parse_args()
 
         item = ItemModel.find_by_name(name)
 
@@ -56,7 +60,7 @@ class Items(Resource):
         # return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
 
     def post(self):
-        data = Item.parser.parse_args()
+        data = _item_parser.parse_args()
 
         if ItemModel.find_by_name(data['name']):
             return {'message': "An item with name '{}' already exists".format(data['name'])}, 400
